Renamed UtilNightMode to UntilTheme
The new title accurately reflects the util class' function.
