+++ app/src/main/java/com/precopia/david/lists/common/buildlogic/AppComponent.kt
@@ -2,7 +2,7 @@ package com.precopia.david.lists.common.buildlogic
 
 import android.app.Application
 import android.content.SharedPreferences
-import com.precopia.david.lists.util.IUtilNightModeContract
+import com.precopia.david.lists.util.IUtilThemeContract
 import com.precopia.domain.repository.IRepositoryContract
 import dagger.BindsInstance
 import dagger.Component
@@ -15,7 +15,7 @@ import javax.inject.Singleton
     UtilNightModeModule::class
 ])
 interface AppComponent {
-    fun utilNightMode(): IUtilNightModeContract
+    fun utilNightMode(): IUtilThemeContract
 
     fun sharedPrefs(): SharedPreferences
