+++ app/src/main/java/com/precopia/david/lists/common/buildlogic/UtilNightModeModule.kt
@@ -3,8 +3,8 @@ package com.precopia.david.lists.common.buildlogic
 import android.app.Application
 import android.content.SharedPreferences
 import com.precopia.david.lists.R
-import com.precopia.david.lists.util.IUtilNightModeContract
-import com.precopia.david.lists.util.UtilNightMode
+import com.precopia.david.lists.util.IUtilThemeContract
+import com.precopia.david.lists.util.UtilTheme
 import dagger.Module
 import dagger.Provides
 import javax.inject.Singleton
@@ -14,8 +14,8 @@ class UtilNightModeModule {
     @Singleton
     @Provides
     fun utilNightMode(application: Application,
-                      sharedPrefs: SharedPreferences): IUtilNightModeContract {
-        return UtilNightMode(
+                      sharedPrefs: SharedPreferences): IUtilThemeContract {
+        return UtilTheme(
                 sharedPrefs,
                 application,
                 application.getString(R.string.night_mode_shared_pref_key)
