+++ app/src/main/java/com/precopia/david/lists/util/IUtilThemeContract.kt
@@ -1,6 +1,6 @@
 package com.precopia.david.lists.util
 
-interface IUtilNightModeContract {
+interface IUtilThemeContract {
     fun isNightModeEnabled(): Boolean
 
     fun setDay()
