+++ app/src/main/java/com/precopia/david/lists/view/preferences/buildlogic/PreferencesLogicFactory.kt
@@ -2,18 +2,18 @@ package com.precopia.david.lists.view.preferences.buildlogic
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.precopia.david.lists.util.IUtilNightModeContract
+import com.precopia.david.lists.util.IUtilThemeContract
 import com.precopia.david.lists.view.preferences.IPreferencesViewContract
 import com.precopia.david.lists.view.preferences.PreferencesLogic
 import com.precopia.domain.repository.IRepositoryContract
 
 class PreferencesLogicFactory(
         private val viewModel: IPreferencesViewContract.ViewModel,
-        private val utilNightMode: IUtilNightModeContract,
+        private val utilTheme: IUtilThemeContract,
         private val userRepo: IRepositoryContract.UserRepository
 ) : ViewModelProvider.NewInstanceFactory() {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        return PreferencesLogic(viewModel, utilNightMode, userRepo) as T
+        return PreferencesLogic(viewModel, utilTheme, userRepo) as T
     }
 }
