+++ app/src/main/java/com/precopia/david/lists/view/preferences/buildlogic/PreferencesModule.kt
@@ -3,7 +3,7 @@ package com.precopia.david.lists.view.preferences.buildlogic
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.precopia.david.lists.common.buildlogic.ViewScope
-import com.precopia.david.lists.util.IUtilNightModeContract
+import com.precopia.david.lists.util.IUtilThemeContract
 import com.precopia.david.lists.view.preferences.IPreferencesViewContract
 import com.precopia.david.lists.view.preferences.PreferencesLogic
 import com.precopia.david.lists.view.preferences.PreferencesViewModel
@@ -23,9 +23,9 @@ class PreferencesModule {
     @ViewScope
     @Provides
     fun factory(viewModel: IPreferencesViewContract.ViewModel,
-                utilNightMode: IUtilNightModeContract,
+                utilTheme: IUtilThemeContract,
                 userRepo: IRepositoryContract.UserRepository): ViewModelProvider.NewInstanceFactory {
-        return PreferencesLogicFactory(viewModel, utilNightMode, userRepo)
+        return PreferencesLogicFactory(viewModel, utilTheme, userRepo)
     }
 
     @JvmSuppressWildcards
